Move analyzer.rb to analyze directory
